requirements.txt
cmudict
cn2an
g2p_en
GPUtil
gradio
jaconv
jieba
langid
librosa
loguru
matplotlib
mecab-python3
num2words
numba
numpy
psutil
pyannote.audio>=3.1.0
pyopenjtalk-prebuilt
pypinyin
PyYAML
requests
scipy
tensorboard
transformers